refactor(framework): move consumption-limit validation out of the registry
The registry is a persistence module, but it encoded the consumption domain's
rules: the three window keys, the enabled/percent shape, the 0-100 clamp, and
the default-merge. That knowledge belongs beside ConsumptionLimit and
DEFAULT_CONSUMPTION_LIMITS, so move sanitizeConsumptionLimit(s) into
consumption.ts; the registry imports the one it needs.

resolveConsumptionLimits stays in the registry (it bridges the persisted
Preferences shape to the domain), but its per-field fallback was dead:
sanitizeConsumptionLimits already fills every window, so a stored value is whole
and the only gap left to default is an absent one. Collapse it to that.
