+++ packages/framework/src/registry.ts
@@ -1,5 +1,5 @@
 import { basename, dirname, join, resolve } from 'node:path'
-import { DEFAULT_CONSUMPTION_LIMITS, type ConsumptionLimit, type ConsumptionLimits } from './consumption.js'
+import { DEFAULT_CONSUMPTION_LIMITS, sanitizeConsumptionLimits, type ConsumptionLimits } from './consumption.js'
 import { nodeFs } from './node-fs.js'
 
 /**
@@ -145,39 +145,6 @@ const PREFERENCE_KEYS = [
   'browser',
 ] as const
 
-const CONSUMPTION_LIMIT_KEYS = ['daily', 'fiveHour', 'session'] as const
-
-/**
- * Read one limit out of a hand-edited or browser-supplied object.
- *
- * `undefined` for anything we can't trust, so the caller falls back to the
- * default rather than to an unguarded account. The percentage is clamped rather
- * than rejected: a plausible-but-out-of-range number is a slip, and honouring
- * the nearest legal value beats silently reverting to something else entirely.
- */
-function sanitizeConsumptionLimit(value: unknown): ConsumptionLimit | undefined {
-  if (typeof value !== 'object' || value === null) return undefined
-  const input = value as Record<string, unknown>
-  const percent = input['percent']
-  if (typeof input['enabled'] !== 'boolean') return undefined
-  if (typeof percent !== 'number' || !Number.isFinite(percent)) return undefined
-  return { enabled: input['enabled'], percent: Math.min(100, Math.max(0, percent)) }
-}
-
-/** Read the three limits, falling back per-limit so one bad entry can't unguard the rest. */
-function sanitizeConsumptionLimits(value: unknown): ConsumptionLimits | undefined {
-  if (typeof value !== 'object' || value === null) return undefined
-  const input = value as Record<string, unknown>
-  const limits = {} as ConsumptionLimits
-  let any = false
-  for (const key of CONSUMPTION_LIMIT_KEYS) {
-    const limit = sanitizeConsumptionLimit(input[key])
-    if (limit) any = true
-    limits[key] = limit ?? DEFAULT_CONSUMPTION_LIMITS[key]
-  }
-  return any ? limits : undefined
-}
-
 /** Keep only the known preference fields, so a hand-edited or browser-supplied
  * object never lands junk (or the wrong type) in the user's home file. */
 function sanitizePreferences(value: unknown): Preferences {
@@ -200,13 +167,9 @@ function sanitizePreferences(value: unknown): Preferences {
  * the settings is still guarded (#519).
  */
 export function resolveConsumptionLimits(preferences: Preferences | undefined): ConsumptionLimits {
-  const set = preferences?.consumptionLimits
-  if (!set) return DEFAULT_CONSUMPTION_LIMITS
-  return {
-    daily: set.daily ?? DEFAULT_CONSUMPTION_LIMITS.daily,
-    fiveHour: set.fiveHour ?? DEFAULT_CONSUMPTION_LIMITS.fiveHour,
-    session: set.session ?? DEFAULT_CONSUMPTION_LIMITS.session,
-  }
+  // sanitizeConsumptionLimits already fills every window, so a stored value is
+  // whole: the only gap left to default is an absent one.
+  return preferences?.consumptionLimits ?? DEFAULT_CONSUMPTION_LIMITS
 }
 
 /**
